refactor(profile): extract form components and fix review issues
- Extract PersonalInfoForm, PasswordForm, FeaturesForm, PreferencesForm
  into src/components/profile/ (profile page.tsx: 925 → 228 lines)
- Make preview date dynamic using new Date() instead of hardcoded value
- Add /preferences redirect to /profile for bookmarked URLs
