+++ src/app/(app)/profile/page.tsx
@@ -3,7 +3,7 @@
 import { useState, useEffect, useCallback } from "react";
 import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
-import { User, Lock, Check, Loader2, Sparkles, ScanLine, Rows3, Target, Tag, Settings2, Calendar, Type } from "lucide-react";
+import { User, Lock, Sparkles, Settings2 } from "lucide-react";
 import {
   updateProfileSchema,
   changePasswordSchema,
@@ -12,51 +12,10 @@ import {
 } from "@/lib/validations";
 import { cn } from "@/lib/utils";
 import { useUser } from "@/components/user-provider";
-
-const CURRENCIES = [
-  { value: "PHP", label: "PHP - Philippine Peso" },
-  { value: "USD", label: "USD - US Dollar" },
-  { value: "EUR", label: "EUR - Euro" },
-  { value: "GBP", label: "GBP - British Pound" },
-  { value: "JPY", label: "JPY - Japanese Yen" },
-  { value: "AUD", label: "AUD - Australian Dollar" },
-  { value: "CAD", label: "CAD - Canadian Dollar" },
-  { value: "SGD", label: "SGD - Singapore Dollar" },
-  { value: "KRW", label: "KRW - South Korean Won" },
-  { value: "INR", label: "INR - Indian Rupee" },
-];
-
-const TIMEZONES = [
-  { value: 720, label: "UTC-12:00" },
-  { value: 660, label: "UTC-11:00" },
-  { value: 600, label: "UTC-10:00 (Honolulu)" },
-  { value: 540, label: "UTC-09:00 (Anchorage)" },
-  { value: 480, label: "UTC-08:00 (Los Angeles)" },
-  { value: 420, label: "UTC-07:00 (Denver)" },
-  { value: 360, label: "UTC-06:00 (Chicago)" },
-  { value: 300, label: "UTC-05:00 (New York)" },
-  { value: 240, label: "UTC-04:00 (Halifax)" },
-  { value: 180, label: "UTC-03:00 (S\u00e3o Paulo)" },
-  { value: 120, label: "UTC-02:00" },
-  { value: 60, label: "UTC-01:00" },
-  { value: 0, label: "UTC+00:00 (London)" },
-  { value: -60, label: "UTC+01:00 (Paris)" },
-  { value: -120, label: "UTC+02:00 (Cairo)" },
-  { value: -180, label: "UTC+03:00 (Moscow)" },
-  { value: -240, label: "UTC+04:00 (Dubai)" },
-  { value: -300, label: "UTC+05:00 (Karachi)" },
-  { value: -330, label: "UTC+05:30 (Mumbai)" },
-  { value: -345, label: "UTC+05:45 (Kathmandu)" },
-  { value: -360, label: "UTC+06:00 (Dhaka)" },
-  { value: -420, label: "UTC+07:00 (Bangkok)" },
-  { value: -480, label: "UTC+08:00 (Manila)" },
-  { value: -540, label: "UTC+09:00 (Tokyo)" },
-  { value: -570, label: "UTC+09:30 (Adelaide)" },
-  { value: -600, label: "UTC+10:00 (Sydney)" },
-  { value: -660, label: "UTC+11:00" },
-  { value: -720, label: "UTC+12:00 (Auckland)" },
-  { value: -780, label: "UTC+13:00 (Samoa)" },
-];
+import { PersonalInfoForm } from "@/components/profile/personal-info-form";
+import { PasswordForm } from "@/components/profile/password-form";
+import { FeaturesForm } from "@/components/profile/features-form";
+import { PreferencesForm } from "@/components/profile/preferences-form";
 
 type Tab = "personal" | "password" | "features" | "preferences";
 
@@ -123,7 +82,6 @@ export default function ProfilePage() {
         throw new Error(err.error || "Failed to update profile");
       }
 
-      // Update the shared user context so sidebar + currency + timezone reflect immediately
       setUser({ name: data.name, email: data.email, currency: data.currency, timezoneOffset: data.timezoneOffset });
 
       setProfileSuccess("Profile updated successfully");
@@ -264,660 +222,6 @@ export default function ProfilePage() {
             <PreferencesForm />
           </div>
         </div>
-      </div>
-    </div>
-  );
-}
-
-const INPUT_CLASS =
-  "w-full px-4 py-3 rounded-xl border border-cream-300 bg-cream-50/50 text-warm-700 placeholder:text-warm-300 focus:outline-none focus:ring-2 focus:ring-amber/30 focus:border-amber transition-all";
-
-interface PersonalInfoFormProps {
-  form: ReturnType<typeof useForm<UpdateProfileInput>>;
-  onSubmit: (data: UpdateProfileInput) => Promise<void>;
-  success: string;
-  error: string;
-}
-
-function PersonalInfoForm({ form, onSubmit, success, error }: PersonalInfoFormProps) {
-  const {
-    register,
-    handleSubmit,
-    formState: { errors, isSubmitting },
-  } = form;
-
-  return (
-    <div className="card p-6">
-      <div className="mb-5">
-        <h2 className="font-serif text-lg text-warm-700">Personal Information</h2>
-        <p className="text-sm text-warm-400 mt-0.5">
-          Update your name, email, currency, and timezone
-        </p>
-      </div>
-
-      <form onSubmit={handleSubmit(onSubmit)} className="space-y-4">
-        <div>
-          <label className="block text-sm font-medium text-warm-600 mb-1.5">
-            Name
-          </label>
-          <input
-            type="text"
-            {...register("name")}
-            className={INPUT_CLASS}
-            placeholder="Your name"
-          />
-          {errors.name && (
-            <p className="text-expense text-sm mt-1">{errors.name.message}</p>
-          )}
-        </div>
-
-        <div>
-          <label className="block text-sm font-medium text-warm-600 mb-1.5">
-            Email
-          </label>
-          <input
-            type="email"
-            {...register("email")}
-            className={INPUT_CLASS}
-            placeholder="your@email.com"
-          />
-          {errors.email && (
-            <p className="text-expense text-sm mt-1">{errors.email.message}</p>
-          )}
-        </div>
-
-        <div>
-          <label className="block text-sm font-medium text-warm-600 mb-1.5">
-            Currency
-          </label>
-          <select
-            {...register("currency")}
-            className={INPUT_CLASS}
-          >
-            {CURRENCIES.map((c) => (
-              <option key={c.value} value={c.value}>
-                {c.label}
-              </option>
-            ))}
-          </select>
-          {errors.currency && (
-            <p className="text-expense text-sm mt-1">{errors.currency.message}</p>
-          )}
-        </div>
-
-        <div>
-          <label className="block text-sm font-medium text-warm-600 mb-1.5">
-            Timezone
-          </label>
-          <select
-            {...register("timezoneOffset", { valueAsNumber: true })}
-            className={INPUT_CLASS}
-          >
-            {TIMEZONES.map((tz) => (
-              <option key={tz.value} value={tz.value}>
-                {tz.label}
-              </option>
-            ))}
-          </select>
-        </div>
-
-        {success && (
-          <div className="flex items-center gap-2 p-3 rounded-xl bg-income-light border border-income/20 text-income-dark text-sm">
-            <Check className="w-4 h-4" />
-            {success}
-          </div>
-        )}
-
-        {error && (
-          <div className="p-3 rounded-xl bg-expense-light border border-expense/20 text-expense-dark text-sm">
-            {error}
-          </div>
-        )}
-
-        <div className="pt-2">
-          <button
-            type="submit"
-            disabled={isSubmitting}
-            className="inline-flex items-center justify-center gap-2 px-5 py-2.5 rounded-xl bg-amber hover:bg-amber-dark text-white font-medium text-sm transition-colors shadow-soft disabled:opacity-50 disabled:cursor-not-allowed"
-          >
-            {isSubmitting ? (
-              <Loader2 className="w-4 h-4 animate-spin" />
-            ) : (
-              "Save Changes"
-            )}
-          </button>
-        </div>
-      </form>
-    </div>
-  );
-}
-
-interface PasswordFormProps {
-  form: ReturnType<typeof useForm<ChangePasswordInput>>;
-  onSubmit: (data: ChangePasswordInput) => Promise<void>;
-  success: string;
-  error: string;
-}
-
-function PasswordForm({ form, onSubmit, success, error }: PasswordFormProps) {
-  const {
-    register,
-    handleSubmit,
-    formState: { errors, isSubmitting },
-  } = form;
-
-  return (
-    <div className="card p-6">
-      <div className="mb-5">
-        <h2 className="font-serif text-lg text-warm-700">Change Password</h2>
-        <p className="text-sm text-warm-400 mt-0.5">
-          Update your password to keep your account secure
-        </p>
-      </div>
-
-      <form onSubmit={handleSubmit(onSubmit)} className="space-y-4">
-        <div>
-          <label className="block text-sm font-medium text-warm-600 mb-1.5">
-            Current Password
-          </label>
-          <input
-            type="password"
-            {...register("currentPassword")}
-            className={INPUT_CLASS}
-            placeholder="Enter current password"
-          />
-          {errors.currentPassword && (
-            <p className="text-expense text-sm mt-1">
-              {errors.currentPassword.message}
-            </p>
-          )}
-        </div>
-
-        <div>
-          <label className="block text-sm font-medium text-warm-600 mb-1.5">
-            New Password
-          </label>
-          <input
-            type="password"
-            {...register("newPassword")}
-            className={INPUT_CLASS}
-            placeholder="Enter new password"
-          />
-          {errors.newPassword && (
-            <p className="text-expense text-sm mt-1">
-              {errors.newPassword.message}
-            </p>
-          )}
-        </div>
-
-        <div>
-          <label className="block text-sm font-medium text-warm-600 mb-1.5">
-            Confirm New Password
-          </label>
-          <input
-            type="password"
-            {...register("confirmPassword")}
-            className={INPUT_CLASS}
-            placeholder="Confirm new password"
-          />
-          {errors.confirmPassword && (
-            <p className="text-expense text-sm mt-1">
-              {errors.confirmPassword.message}
-            </p>
-          )}
-        </div>
-
-        {success && (
-          <div className="flex items-center gap-2 p-3 rounded-xl bg-income-light border border-income/20 text-income-dark text-sm">
-            <Check className="w-4 h-4" />
-            {success}
-          </div>
-        )}
-
-        {error && (
-          <div className="p-3 rounded-xl bg-expense-light border border-expense/20 text-expense-dark text-sm">
-            {error}
-          </div>
-        )}
-
-        <div className="pt-2">
-          <button
-            type="submit"
-            disabled={isSubmitting}
-            className="inline-flex items-center justify-center gap-2 px-5 py-2.5 rounded-xl bg-amber hover:bg-amber-dark text-white font-medium text-sm transition-colors shadow-soft disabled:opacity-50 disabled:cursor-not-allowed"
-          >
-            {isSubmitting ? (
-              <Loader2 className="w-4 h-4 animate-spin" />
-            ) : (
-              "Change Password"
-            )}
-          </button>
-        </div>
-      </form>
-    </div>
-  );
-}
-
-function FeaturesForm() {
-  const { user, setUser } = useUser();
-  const [saving, setSaving] = useState(false);
-  const [savingLayout, setSavingLayout] = useState(false);
-  const [savingAutofocus, setSavingAutofocus] = useState(false);
-  const [savingLabelType, setSavingLabelType] = useState(false);
-
-  const handleToggle = async () => {
-    const newValue = !user.receiptScanEnabled;
-
-    // Optimistic update for instant nav response
-    setUser({ receiptScanEnabled: newValue });
-    setSaving(true);
-
-    try {
-      const res = await fetch("/api/preferences", {
-        method: "PATCH",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ receiptScanEnabled: newValue }),
-      });
-
-      if (!res.ok) {
-        // Revert on failure
-        setUser({ receiptScanEnabled: !newValue });
-      }
-    } catch {
-      // Revert on network error
-      setUser({ receiptScanEnabled: !newValue });
-    } finally {
-      setSaving(false);
-    }
-  };
-
-  const handleLayoutToggle = async () => {
-    const newValue = user.transactionLayout === "infinite" ? "pagination" : "infinite";
-    const oldValue = user.transactionLayout;
-
-    setUser({ transactionLayout: newValue });
-    setSavingLayout(true);
-
-    try {
-      const res = await fetch("/api/preferences", {
-        method: "PATCH",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ transactionLayout: newValue }),
-      });
-
-      if (!res.ok) {
-        setUser({ transactionLayout: oldValue });
-      }
-    } catch {
-      setUser({ transactionLayout: oldValue });
-    } finally {
-      setSavingLayout(false);
-    }
-  };
-
-  const handleAutofocusToggle = async () => {
-    const newValue = !user.transactionAmountAutofocus;
-    const oldValue = user.transactionAmountAutofocus;
-
-    setUser({ transactionAmountAutofocus: newValue });
-    setSavingAutofocus(true);
-
-    try {
-      const res = await fetch("/api/preferences", {
-        method: "PATCH",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ transactionAmountAutofocus: newValue }),
-      });
-
-      if (!res.ok) {
-        setUser({ transactionAmountAutofocus: oldValue });
-      }
-    } catch {
-      setUser({ transactionAmountAutofocus: oldValue });
-    } finally {
-      setSavingAutofocus(false);
-    }
-  };
-
-  const handleLabelTypeChange = async (newValue: "EXPENSE" | "INCOME" | "BOTH") => {
-    const oldValue = user.defaultLabelType;
-    if (newValue === oldValue) return;
-
-    setUser({ defaultLabelType: newValue });
-    setSavingLabelType(true);
-
-    try {
-      const res = await fetch("/api/preferences", {
-        method: "PATCH",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ defaultLabelType: newValue }),
-      });
-
-      if (!res.ok) {
-        setUser({ defaultLabelType: oldValue });
-      }
-    } catch {
-      setUser({ defaultLabelType: oldValue });
-    } finally {
-      setSavingLabelType(false);
-    }
-  };
-
-  return (
-    <div className="card p-6">
-      <div className="mb-5">
-        <h2 className="font-serif text-lg text-warm-700">Features</h2>
-        <p className="text-sm text-warm-400 mt-0.5">
-          Enable or disable optional features
-        </p>
-      </div>
-
-      <div className="space-y-4">
-        <div className="flex items-center justify-between gap-4 p-4 rounded-xl border border-cream-300 bg-cream-50/50">
-          <div className="flex items-center gap-3">
-            <div className="w-10 h-10 rounded-xl bg-amber-light flex items-center justify-center">
-              <ScanLine className="w-5 h-5 text-amber-dark" />
-            </div>
-            <div>
-              <p className="text-sm font-medium text-warm-600">
-                Receipt Scanning
-              </p>
-              <p className="text-xs text-warm-400">
-                Add a Scan button to mobile navigation for capturing receipts
-              </p>
-            </div>
-          </div>
-
-          {user.roleScanEnabled ? (
-            <button
-              type="button"
-              role="switch"
-              aria-checked={user.receiptScanEnabled}
-              disabled={saving}
-              onClick={handleToggle}
-              className={cn(
-                "relative inline-flex h-6 w-11 shrink-0 cursor-pointer rounded-full border-2 border-transparent transition-colors duration-200 focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-amber/30 disabled:opacity-50 disabled:cursor-not-allowed",
-                user.receiptScanEnabled ? "bg-amber" : "bg-cream-300"
-              )}
-            >
-              <span
-                className={cn(
-                  "pointer-events-none inline-block h-5 w-5 rounded-full bg-white shadow-sm transition-transform duration-200",
-                  user.receiptScanEnabled ? "translate-x-5" : "translate-x-0"
-                )}
-              />
-            </button>
-          ) : (
-            <span className="text-xs text-warm-400 bg-cream-200 px-3 py-1 rounded-full">
-              Not available
-            </span>
-          )}
-        </div>
-
-        <div className="flex items-center justify-between gap-4 p-4 rounded-xl border border-cream-300 bg-cream-50/50">
-          <div className="flex items-center gap-3">
-            <div className="w-10 h-10 rounded-xl bg-amber-light flex items-center justify-center">
-              <Rows3 className="w-5 h-5 text-amber-dark" />
-            </div>
-            <div>
-              <p className="text-sm font-medium text-warm-600">
-                Infinite Scroll
-              </p>
-              <p className="text-xs text-warm-400">
-                Load transactions as you scroll instead of using page navigation
-              </p>
-            </div>
-          </div>
-
-          <button
-            type="button"
-            role="switch"
-            aria-checked={user.transactionLayout === "infinite"}
-            disabled={savingLayout}
-            onClick={handleLayoutToggle}
-            className={cn(
-              "relative inline-flex h-6 w-11 shrink-0 cursor-pointer rounded-full border-2 border-transparent transition-colors duration-200 focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-amber/30 disabled:opacity-50 disabled:cursor-not-allowed",
-              user.transactionLayout === "infinite" ? "bg-amber" : "bg-cream-300"
-            )}
-          >
-            <span
-              className={cn(
-                "pointer-events-none inline-block h-5 w-5 rounded-full bg-white shadow-sm transition-transform duration-200",
-                user.transactionLayout === "infinite" ? "translate-x-5" : "translate-x-0"
-              )}
-            />
-          </button>
-        </div>
-
-        <div className="flex items-center justify-between gap-4 p-4 rounded-xl border border-cream-300 bg-cream-50/50">
-          <div className="flex items-center gap-3">
-            <div className="w-10 h-10 rounded-xl bg-amber-light flex items-center justify-center">
-              <Target className="w-5 h-5 text-amber-dark" />
-            </div>
-            <div>
-              <p className="text-sm font-medium text-warm-600">
-                Auto-focus Amount
-              </p>
-              <p className="text-xs text-warm-400">
-                Focus Amount field first when opening a new transaction form
-              </p>
-            </div>
-          </div>
-
-          <button
-            type="button"
-            role="switch"
-            aria-checked={user.transactionAmountAutofocus}
-            disabled={savingAutofocus}
-            onClick={handleAutofocusToggle}
-            className={cn(
-              "relative inline-flex h-6 w-11 shrink-0 cursor-pointer rounded-full border-2 border-transparent transition-colors duration-200 focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-amber/30 disabled:opacity-50 disabled:cursor-not-allowed",
-              user.transactionAmountAutofocus ? "bg-amber" : "bg-cream-300"
-            )}
-          >
-            <span
-              className={cn(
-                "pointer-events-none inline-block h-5 w-5 rounded-full bg-white shadow-sm transition-transform duration-200",
-                user.transactionAmountAutofocus ? "translate-x-5" : "translate-x-0"
-              )}
-            />
-          </button>
-        </div>
-
-        <div className="p-4 rounded-xl border border-cream-300 bg-cream-50/50">
-          <div className="flex items-center gap-3 mb-3">
-            <div className="w-10 h-10 rounded-xl bg-amber-light flex items-center justify-center">
-              <Tag className="w-5 h-5 text-amber-dark" />
-            </div>
-            <div>
-              <p className="text-sm font-medium text-warm-600">
-                Default Label Type
-              </p>
-              <p className="text-xs text-warm-400">
-                Default transaction type restriction for newly created labels
-              </p>
-            </div>
-          </div>
-
-          <div className="flex gap-1.5 p-1 bg-cream-100 rounded-xl">
-            {(["EXPENSE", "INCOME", "BOTH"] as const).map((type) => (
-              <button
-                key={type}
-                type="button"
-                disabled={savingLabelType}
-                onClick={() => handleLabelTypeChange(type)}
-                className={cn(
-                  "flex-1 py-2 rounded-lg text-xs font-medium transition-all duration-200 disabled:opacity-50",
-                  user.defaultLabelType === type
-                    ? "bg-white text-warm-700 shadow-warm"
-                    : "text-warm-400 hover:text-warm-600"
-                )}
-              >
-                {type === "EXPENSE" ? "Expense" : type === "INCOME" ? "Income" : "Both"}
-              </button>
-            ))}
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-const DAY_FORMAT_OPTIONS = [
-  { value: "SHORT" as const, label: "Short", example: "Mon" },
-  { value: "FULL" as const, label: "Full", example: "Monday" },
-];
-
-function PreferencesForm() {
-  const { user, setUser } = useUser();
-  const [savingDayName, setSavingDayName] = useState(false);
-  const [savingDayFormat, setSavingDayFormat] = useState(false);
-
-  const handleDayNameToggle = async () => {
-    const newValue = !user.showDayName;
-    const oldValue = user.showDayName;
-
-    setUser({ showDayName: newValue });
-    setSavingDayName(true);
-
-    try {
-      const res = await fetch("/api/preferences", {
-        method: "PATCH",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ showDayName: newValue }),
-      });
-
-      if (!res.ok) {
-        setUser({ showDayName: oldValue });
-      }
-    } catch {
-      setUser({ showDayName: oldValue });
-    } finally {
-      setSavingDayName(false);
-    }
-  };
-
-  const handleDayFormatChange = async (newValue: "FULL" | "SHORT") => {
-    const oldValue = user.dayNameFormat;
-    if (newValue === oldValue) return;
-
-    setUser({ dayNameFormat: newValue });
-    setSavingDayFormat(true);
-
-    try {
-      const res = await fetch("/api/preferences", {
-        method: "PATCH",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ dayNameFormat: newValue }),
-      });
-
-      if (!res.ok) {
-        setUser({ dayNameFormat: oldValue });
-      }
-    } catch {
-      setUser({ dayNameFormat: oldValue });
-    } finally {
-      setSavingDayFormat(false);
-    }
-  };
-
-  return (
-    <div className="card p-6">
-      <div className="mb-5">
-        <h2 className="font-serif text-lg text-warm-700">Preferences</h2>
-        <p className="text-sm text-warm-400 mt-0.5">
-          Date headers and formatting options
-        </p>
-      </div>
-
-      <div className="space-y-4">
-        {/* Show Day Name toggle */}
-        <div className="flex items-center justify-between gap-4 p-4 rounded-xl border border-cream-300 bg-cream-50/50">
-          <div className="flex items-center gap-3">
-            <div className="w-10 h-10 rounded-xl bg-amber-light flex items-center justify-center">
-              <Calendar className="w-5 h-5 text-amber-dark" />
-            </div>
-            <div>
-              <p className="text-sm font-medium text-warm-600">
-                Show Day Name
-              </p>
-              <p className="text-xs text-warm-400">
-                Display the day of the week on date headers
-              </p>
-            </div>
-          </div>
-
-          <button
-            type="button"
-            role="switch"
-            aria-checked={user.showDayName}
-            disabled={savingDayName}
-            onClick={handleDayNameToggle}
-            className={cn(
-              "relative inline-flex h-6 w-11 shrink-0 cursor-pointer rounded-full border-2 border-transparent transition-colors duration-200 focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-amber/30 disabled:opacity-50 disabled:cursor-not-allowed",
-              user.showDayName ? "bg-amber" : "bg-cream-300"
-            )}
-          >
-            <span
-              className={cn(
-                "pointer-events-none inline-block h-5 w-5 rounded-full bg-white shadow-sm transition-transform duration-200",
-                user.showDayName ? "translate-x-5" : "translate-x-0"
-              )}
-            />
-          </button>
-        </div>
-
-        {/* Day Name Format */}
-        {user.showDayName && (
-          <div className="p-4 rounded-xl border border-cream-300 bg-cream-50/50">
-            <div className="flex items-center gap-3 mb-3">
-              <div className="w-10 h-10 rounded-xl bg-amber-light flex items-center justify-center">
-                <Type className="w-5 h-5 text-amber-dark" />
-              </div>
-              <div>
-                <p className="text-sm font-medium text-warm-600">
-                  Day Name Format
-                </p>
-                <p className="text-xs text-warm-400">
-                  Choose between short or full day names
-                </p>
-              </div>
-            </div>
-
-            <div className="flex gap-2 ml-[52px]">
-              {DAY_FORMAT_OPTIONS.map((option) => (
-                <button
-                  key={option.value}
-                  onClick={() => handleDayFormatChange(option.value)}
-                  disabled={savingDayFormat}
-                  className={cn(
-                    "px-4 py-2 rounded-lg text-sm font-medium transition-all duration-200 disabled:opacity-50",
-                    user.dayNameFormat === option.value
-                      ? "bg-amber text-white shadow-soft"
-                      : "bg-cream-200 text-warm-500 hover:bg-cream-300"
-                  )}
-                >
-                  {option.example}
-                </button>
-              ))}
-            </div>
-          </div>
-        )}
-
-        {/* Preview */}
-        {user.showDayName && (
-          <div className="p-4 rounded-xl border border-dashed border-cream-300 bg-cream-50/30">
-            <p className="text-[10px] text-warm-300 uppercase tracking-wider mb-2">Preview</p>
-            <div className="flex items-center gap-2 px-4 py-2.5 bg-cream-50 border border-cream-200 rounded-lg">
-              <span className="text-xs font-semibold text-warm-500 uppercase tracking-wide">
-                April 6, 2026
-                <span className="text-warm-300 font-normal normal-case tracking-normal ml-1.5">
-                  · {user.dayNameFormat === "FULL" ? "Sunday" : "Sun"}
-                </span>
-              </span>
-            </div>
-          </div>
-        )}
       </div>
     </div>
   );
